+++ aws-android-sdk-cognitoidentityprovider/src/main/java/com/amazonaws/mobileconnectors/cognitoidentityprovider/CognitoUserSession.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2013-2016 Amazon.com,
+ *  Copyright 2013-2019 Amazon.com,
  *  Inc. or its affiliates. All Rights Reserved.
  *
  *  Licensed under the Amazon Software License (the "License").
@@ -17,6 +17,8 @@
 
 package com.amazonaws.mobileconnectors.cognitoidentityprovider;
 
+import android.util.Log;
+
 import com.amazonaws.SDKGlobalConfiguration;
 import com.amazonaws.mobileconnectors.cognitoidentityprovider.tokens.CognitoAccessToken;
 import com.amazonaws.mobileconnectors.cognitoidentityprovider.tokens.CognitoIdToken;
@@ -24,12 +26,15 @@ import com.amazonaws.mobileconnectors.cognitoidentityprovider.tokens.CognitoRefr
 import com.amazonaws.mobileconnectors.cognitoidentityprovider.util.CognitoIdentityProviderClientConfig;
 
 import java.util.Date;
+import java.util.Objects;
 
 /**
  * This wraps all Cognito tokens for a user.
  */
 public class CognitoUserSession {
     private static final int SECS_CONVERSION = 1000;
+
+    private static final String TAG = CognitoUserSession.class.getSimpleName();
 
     /**
      * Cognito identity token.
@@ -95,6 +100,16 @@ public class CognitoUserSession {
         final Date currentTimeStamp = new Date();
 
         try {
+            if (idToken == null) {
+                Log.w(TAG, "CognitoUserSession is not valid because idToken is null.");
+                return false;
+            }
+
+            if (accessToken == null) {
+                Log.w(TAG, "CognitoUserSession is not valid because accessToken is null.");
+                return false;
+            }
+
             return (currentTimeStamp.before(idToken.getExpiration())
                     & currentTimeStamp.before(accessToken.getExpiration()));
         } catch (final Exception e) {
@@ -103,16 +118,36 @@ public class CognitoUserSession {
     }
 
     /**
-     * Returns true if this session for the threshold set in {@link CognitoIdentityProviderClientConfig#refreshThreshold}.
+     * Returns true if this session is valid for the threshold set in
+     * {@link CognitoIdentityProviderClientConfig#getRefreshThreshold}.
      *
-     * @return boolean to indicate if the session is valid for atleast {@link CognitoIdentityProviderClientConfig#refreshThreshold} seconds.
+     * Checks for the following conditions to determine if the session is valid:
+     * 1) Existence of an idToken
+     * 2) Existence of an accessToken
+     * 3) idToken is not expired
+     * 4) accessToken is not expired
+     *
+     * @return boolean to indicate if the session is valid for at-least
+     * {@link CognitoIdentityProviderClientConfig#getRefreshThreshold} seconds.
      */
     public boolean isValidForThreshold() {
         try {
+            if (idToken == null) {
+                Log.w(TAG, "CognitoUserSession is not valid because idToken is null.");
+                return false;
+            }
+
+            if (accessToken == null) {
+                Log.w(TAG, "CognitoUserSession is not valid because accessToken is null.");
+                return false;
+            }
+
             final long currentTime = System.currentTimeMillis()
                     - SDKGlobalConfiguration.getGlobalTimeOffset() * SECS_CONVERSION;
-            final long expiresInMilliSeconds = idToken.getExpiration().getTime() - currentTime;
-            return (expiresInMilliSeconds > CognitoIdentityProviderClientConfig.getRefreshThreshold());
+            final long idTokenExpiresInMilliSeconds = idToken.getExpiration().getTime() - currentTime;
+            final long accessTokenExpiresInMilliSeconds = accessToken.getExpiration().getTime() - currentTime;
+            return (idTokenExpiresInMilliSeconds > CognitoIdentityProviderClientConfig.getRefreshThreshold()) &&
+                    (accessTokenExpiresInMilliSeconds > CognitoIdentityProviderClientConfig.getRefreshThreshold());
         } catch (final Exception e) {
             return false;
         }