+++ CHANGELOG.md
@@ -24,6 +24,20 @@
     - [#1109](https://github.com/aws-amplify/aws-sdk-android/issues/1109)
     - [#1115](https://github.com/aws-amplify/aws-sdk-android/issues/1115)
 
+- **Amazon S3**
+  - Fixed a bug where multi-part uploads via `TransferUtility` would fail to propagate tags to `Amazon S3` from the `UserMetadata` passed through the `ObjectMetadata`. See [Issue#541](https://github.com/aws-amplify/aws-sdk-android/issues/541).
+  - The following code should now attach a tag for both single-part and multi-part uploads:
+  
+	```java
+	ObjectMetadata metadata = new ObjectMetadata();
+	metadata.addUserMetadata(Headers.S3_TAGGING, "key=value");
+	TransferObserver observer = transferUtility.upload(
+		file.getName(),
+		file,
+		metadata
+	);
+	```
+
 ## [Release 2.15.0](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.15.0)
 
 ### Bug Fixes
@@ -44,6 +58,7 @@
   - AWS IoT
   - Amazon Lex
   - Amazon Rekognition
+  - Amazon AutoScaling
 
 ## [Release 2.14.2](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.14.2)